Move Tile struct into a submodule
